Expose the query name for every invoice line item
By returning the query names for line items, ERP adapters can detect
the source of a line item and handle the invoice rendering accordingly.
This can for example be helpful for sub-items.

We may or may not need this for the Odoo adapter, but I don't see a
reason to hide this information