Fix max of (reserved, used) memory
Prometheus returns the first value if the condition is true. No value otherwise.

With the old query we always selected the working_set value since it was the first value in both equations.

Switching the values around to have both in the first position should fix this query.

Old: `vector(1) >= vector(3) or vector(1) < vector(3) == 1`
New: `vector(1) >= vector(3) or vector(3) > vector(1) == 3`